Add files via upload
 Implement Prim’s algorithm using an adjacency matrix.  The text implements Prim’s algorithm using 
lists for adjacent edges.  Implement the algorithm using an adjacency matrix for weighted graphs 
based on the graph provided below.
 Be sure to list the order of the nodes visited, and, for Prim’s always start at Node A.  The distance 
traveled will be the answer we are looking for.  The distance travelled should be the same for both 
Prim’s and Kruskal’s, but the difference between them will be the order of the nodes visited.